Refactor GroupProfile into new accounts app
