+++ psat_server_web/atlas/accounts/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class ATLASAccountsConfig(AppConfig):
+    name = 'accounts'
+
+    # def ready(self) -> None:
+    #     import accounts.signals
