--- psat_server_web/atlas/atlas/apps.py
@@ -1,8 +1,0 @@
-from django.apps import AppConfig
-
-
-class ATLASWebConfig(AppConfig):
-    name = 'atlas'
-
-    def ready(self) -> None:
-        import atlas.signals
