--- psat_server_web/atlas/atlas/signals.py
@@ -1,9 +1,0 @@
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-from django.contrib.auth.models import Group
-from .models import GroupProfile
-
-@receiver(post_save, sender=Group)
-def create_group_profile(sender, instance, created, **kwargs):
-    if created:
-        GroupProfile.objects.create(group=instance)
